application folder changes

+++ application/models/Tasks_model.php
@@ -317,7 +317,7 @@ class Tasks_model extends App_Model
 
     public function get_billable_tasks($customer_id = false, $project_id = '')
     {
-        $has_permission_view = has_permission('tasks', '', 'view');
+        $has_permission_view = staff_can('view',  'tasks');
         $noPermissionsQuery  = get_tasks_where_string(false);
 
         $this->db->where('billable', 1);
@@ -1205,12 +1205,14 @@ class Tasks_model extends App_Model
             if (empty($attachment->external)) {
                 $relPath  = get_upload_path_by_type('task') . $attachment->rel_id . '/';
                 $fullPath = $relPath . $attachment->file_name;
-                unlink($fullPath);
-                $fname     = pathinfo($fullPath, PATHINFO_FILENAME);
-                $fext      = pathinfo($fullPath, PATHINFO_EXTENSION);
-                $thumbPath = $relPath . $fname . '_thumb.' . $fext;
-                if (file_exists($thumbPath)) {
-                    unlink($thumbPath);
+                if(file_exists($fullPath)) {
+                    unlink($fullPath);
+                    $fname     = pathinfo($fullPath, PATHINFO_FILENAME);
+                    $fext      = pathinfo($fullPath, PATHINFO_EXTENSION);
+                    $thumbPath = $relPath . $fname . '_thumb.' . $fext;
+                    if (file_exists($thumbPath)) {
+                        unlink($thumbPath);
+                    }
                 }
             }
 
@@ -1382,7 +1384,7 @@ class Tasks_model extends App_Model
         // Check if user really creator
         $this->db->where('id', $data['id']);
         $comment = $this->db->get(db_prefix() . 'task_comments')->row();
-        if ($comment->staffid == get_staff_user_id() || has_permission('tasks', '', 'edit') || $comment->contact_id == get_contact_user_id()) {
+        if ($comment->staffid == get_staff_user_id() || staff_can('edit',  'tasks') || $comment->contact_id == get_contact_user_id()) {
             $comment_added = strtotime($comment->dateadded);
             $minus_1_hour  = strtotime('-1 hours');
             if (get_option('client_staff_add_edit_delete_task_comments_first_hour') == 0 || (get_option('client_staff_add_edit_delete_task_comments_first_hour') == 1 && $comment_added >= $minus_1_hour) || is_admin()) {
@@ -1425,7 +1427,7 @@ class Tasks_model extends App_Model
             return true;
         }
 
-        if ($comment->staffid == get_staff_user_id() || has_permission('tasks', '', 'delete') || $comment->contact_id == get_contact_user_id() || $force === true) {
+        if ($comment->staffid == get_staff_user_id() || staff_can('delete',  'tasks') || $comment->contact_id == get_contact_user_id() || $force === true) {
             $comment_added = strtotime($comment->dateadded);
             $minus_1_hour  = strtotime('-1 hours');
             if (
@@ -2120,7 +2122,7 @@ class Tasks_model extends App_Model
     {
         $task_id = $this->db->escape_str($task_id);
 
-        return $this->db->query("SELECT id,note,start_time,end_time,task_id,staff_id, CONCAT(firstname, ' ', lastname) as full_name,
+        return $this->db->query("SELECT id," . db_prefix() . "taskstimers.note as note, start_time,end_time,task_id,staff_id, CONCAT(firstname, ' ', lastname) as full_name,
         end_time - start_time time_spent FROM " . db_prefix() . 'taskstimers JOIN ' . db_prefix() . 'staff ON ' . db_prefix() . 'staff.staffid=' . db_prefix() . "taskstimers.staff_id WHERE task_id = '$task_id' ORDER BY start_time DESC")->result_array();
     }
